setting基本结构

+++ background/utils.js
@@ -23,14 +23,23 @@ const setStorage = (key, value) => {
 const http = {
   get: async (url, params = {}) => {
     const xAuthToken = await getStorage("xAuthToken");
-    const result = await axios.get(`https://www.yuque.com/api/v2${url}`, {
-      params,
-      headers: {
-        "Content-Type": "application/json",
-        "X-Auth-Token": xAuthToken,
-      },
-    });
-    if (result.status === 200) {
+    if (!xAuthToken) {
+      alert("请先设置用户token。");
+      return;
+    }
+    const result = await axios
+      .get(`https://www.yuque.com/api/v2${url}`, {
+        params,
+        headers: {
+          "Content-Type": "application/json",
+          "X-Auth-Token": xAuthToken,
+        },
+      })
+      .catch((err) => {
+        console.log("from http.get catch ...", err);
+        alert("token有误或token权限不足，请检查token及权限。");
+      });
+    if (result) {
       // 返回值 result.data 格式为 { data: {} }
       return result.data;
     }
@@ -38,6 +47,10 @@ const http = {
   },
   post: async (url, params = {}) => {
     const xAuthToken = await getStorage("xAuthToken");
+    if (!xAuthToken) {
+      alert("请先设置用户token。");
+      return;
+    }
     const result = await axios.post(
       `https://www.yuque.com/api/v2${url}`,
       params,
@@ -56,6 +69,10 @@ const http = {
   },
   put: async (url, params = {}) => {
     const xAuthToken = await getStorage("xAuthToken");
+    if (!xAuthToken) {
+      alert("请先设置用户token。");
+      return;
+    }
     const result = await axios.put(
       `https://www.yuque.com/api/v2${url}`,
       params,
@@ -74,6 +91,10 @@ const http = {
   },
   delete: async (url) => {
     const xAuthToken = await getStorage("xAuthToken");
+    if (!xAuthToken) {
+      alert("请先设置用户token。");
+      return;
+    }
     const result = await axios.delete(`https://www.yuque.com/api/v2${url}`, {
       headers: {
         "Content-Type": "application/json",
